Moved to strict ipc communication
All requests from renderer to main are send using ipcRender invoke and handle by the main thread using ipcMain handle.

This way the separation between renderer and main thread is secured. Previous implementation of exposing whole main thread classes was not great and lead to a mixup of main and renderer thread. This complicated things way to much.